+++ example/fortnite_api_io.dart
@@ -1,9 +1,8 @@
-import "package:fortnite/fortnite.dart";
-import "package:fortnite/src/fortnite_api_io/structures/challenges.dart";
+import "package:fortnite/fortnite_api_io.dart";
 
 void main() async {
   final FortniteApiIo fn = FortniteApiIo(
-    apiKey: "2840821f-df7a0c5e-891f3bb7-df9aa92a",
+    apiKey: "YOUR API KEY",
   );
 
   ChallengeResponse c = await fn.getChallenges();