+++ lib/fortnite.dart
@@ -13,5 +13,3 @@ export "src/util/utils.dart";
 export "src/client/client.dart";
 export "src/client/http.dart";
 export "src/client/auth.dart";
-
-export "src/fortnite_api_io/fortnite_api_io.dart";