+++ lib/fortnite_api_io.dart
@@ -1,0 +1,4 @@
+export "src/fortnite_api_io/fortnite_api_io.dart";
+
+export "src/fortnite_api_io/structures/challenges.dart";
+export "src/fortnite_api_io/structures/items.dart";