+++ CHANGELOG.md
@@ -36,3 +36,4 @@
 -   Display name getter to client.
 -   FortniteApiIo manager.
 -   Get challenges method in FortniteApiIo class.
+-   Get items method in FortniteApiIo class.